requirements.txt
gymnasium
readchar
pytorch
stable-baselines3[extra]>=2.0.0a4

pygame
pyglet